perf: introduce managers
